🚚 rename: AccountDeletion 페이지 폴더에 components 폴더 생성 후 컴포넌트 이동 #42
