+++ src/pages/AccountDeletion/index.tsx
@@ -1,5 +1,5 @@
 import CryingIcons from "@/assets/heroicons_ticket-solid.svg?react";
-import AccountDeletionForm from "@/pages/AccountDeletion/AccountDeletionForm";
+import AccountDeletionForm from "@/pages/AccountDeletion/components/AccountDeletionForm";
 export default function AccountDeletion() {
   return (
     <>
